bdev: find LVs by name
Ganeti supports LV allocation on different VGs, however it does not
allow moving LVs between VGs in an easy way. Since Ganeti uses UUIDs for
LV names, we can relax the default checks and find an instance's disk by
LV name _only_, disregarding the VG name in the unique_id. This allows
administrators to easily merge/split volume groups.

To this end we introduce LogicalVolume._FindLvPathByName, which uses lvs
--select under the hood to enumerate all LVs matching a given name. In
order to guarantee consistency, this implementation kicks in only if the
regular disk device node does not exist and will fail silently if
any anomaly is detected (e.g. multiple LVs with the same name on
different VGs).

Note that this is only meant as a temporary solution to allow moving LVs
between VGs, so we make sure we emit a warning everytime this mechanism
kicks in. Also note that verify-disks will complain for out-of-VG LVs.