impexpd: fix certificate verification with new socat versions
Socat versions after 1.7.3 verify the server certificate's subject
against either the hostname, or the openssl-commonname option. Since
ganeti uses 'ganeti.example.com' for all self-signed certs, certificate
verification will fail, as socat will be told to connect to the node
using its proper name.

Fix this by passing the openssl-commonname option. Since this option is
only available on newer socat versions and older socat versions will
break when passed the unknown option, we need to parse `socat -V` output
to check if we need to specify the option or not.

This fixes #1226.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>