watcher: change watcher-generated job archival strategy
Currently the watcher spawns a single OP_GROUP_VERIFY_DISKS job per
node-group, which it archives as soon as it has finished (presumably to
avoid polluting the job queue with internal jobs).

Unfortunately this strategy is prone to a race that potentially leaves
stale job files on all master candidates. The race comes down to the
fact that ArchiveJob is called before the job has finished replicating
to all master candidates, as these are two isolated threads of
execution:

 1. The replication is triggered by the job code itself (via
    UpdateJobUnlocked) after each write. The job code runs on a dedicated
    process.

 2. Watcher waits for LuxiD to signal the job's completion, and LuxiD
    does so by watching the job file on the master using inotify. There
    is no locking involved and LuxiD will signal job completion as soon
    as the file has been updated on the master, even though the updated
    file has not hit all master candidates yet.

There are a number of approaches to solving this issue, including:

 - Marking these ephemeral jobs as non-replicatable. While there is a
   provision to skip replicating jobs, it is not exposed to upper layers
   (e.g. via an argument to SubmitJob).

 - Introduce job-level locking between LuxiD and the job code. This is
   probably too complicated and does not seem to be needed anywhere
   else.

Instead, we'll refactor the archival code to cleanup these jobs on the
next watcher run. This is done by marking jobs as "bulk" using the
reason trail and querying the queue for such jobs after they have
finished. While this approach does not eliminate the race, it certainly
reduces the window so that we don't expect any duplicate job files to
emerge in real world scenarios.