Drop Arbitrary instances provided by QuickCheck 2.8
QuickCheck 2.8 provides a Set.Set arbitrary builtin that overlaps with
the one in test/hs/Test/Ganeti/Objects.hs. Adding a CPP check to that
file is not trivial because of special characters in the Python
multiline strings. Converting TagSet to a proper type is not trivial
either.

Instead, let's drop support for QuickCheck versions before 2.8 (and thus
GHC before 7.10) and remove a couple of CPP conditionals.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>