Avoid backing up the export dir on cluster upgrade
The export dir may happen to be under data dir (in fact, this has been
the case with the Debian package since 2.0.1-1). Since the
export dir contains instance disk data, it might have a considerable
size and we don't want it backed up on gnt-cluster upgrade.

This has been carried as a patch in Debian for quite some
time in a much simpler form.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>