Python: use `except … as` everywhere
This is a small step to modernize the Python code. The `except … , x`
syntax has been removed in Python 3 in favor of `except … as x`. The
latter was introduced in Python 2.6, which is our minimum supported
version, thus we can convert all uses of except to the "new" syntax.
While at it, also remove all uses of `except … , _`.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>