vcluster-setup: clobber /etc/hosts instead of renaming
vcluster-setup generates a new /etc/hosts in a temporary and then
renames the latter on top of the former. This works on regular systems
and has the added bonus of replacing /etc/hosts atomically, however it
does not work inside Docker containers or other systems where /etc/hosts
is bind-mounted. Trade some atomicity for functionality and clobber the
existing /etc/hosts file.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>